+++ tools/update_lint_rules/lib/src/models/recommended_rule_severity.freezed.dart
@@ -1,0 +1,484 @@
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
+
+part of 'recommended_rule_severity.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
+/// @nodoc
+mixin _$RecommendedRuleSeverity {
+  Rule get rule => throw _privateConstructorUsedError;
+  String get reason => throw _privateConstructorUsedError;
+  SeverityLevel get severityLevel => throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult when<TResult extends Object?>({
+    required TResult Function(
+            Rule rule, String reason, SeverityLevel severityLevel)
+        dart,
+    required TResult Function(
+            Rule rule, String reason, SeverityLevel severityLevel)
+        flutter,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult? Function(Rule rule, String reason, SeverityLevel severityLevel)?
+        dart,
+    TResult? Function(Rule rule, String reason, SeverityLevel severityLevel)?
+        flutter,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function(Rule rule, String reason, SeverityLevel severityLevel)?
+        dart,
+    TResult Function(Rule rule, String reason, SeverityLevel severityLevel)?
+        flutter,
+    required TResult orElse(),
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult map<TResult extends Object?>({
+    required TResult Function(RecommendedRuleSeverityDart value) dart,
+    required TResult Function(RecommendedRuleSeverityFlutter value) flutter,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult? Function(RecommendedRuleSeverityDart value)? dart,
+    TResult? Function(RecommendedRuleSeverityFlutter value)? flutter,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(RecommendedRuleSeverityDart value)? dart,
+    TResult Function(RecommendedRuleSeverityFlutter value)? flutter,
+    required TResult orElse(),
+  }) =>
+      throw _privateConstructorUsedError;
+
+  @JsonKey(ignore: true)
+  $RecommendedRuleSeverityCopyWith<RecommendedRuleSeverity> get copyWith =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $RecommendedRuleSeverityCopyWith<$Res> {
+  factory $RecommendedRuleSeverityCopyWith(RecommendedRuleSeverity value,
+          $Res Function(RecommendedRuleSeverity) then) =
+      _$RecommendedRuleSeverityCopyWithImpl<$Res, RecommendedRuleSeverity>;
+  @useResult
+  $Res call({Rule rule, String reason, SeverityLevel severityLevel});
+
+  $RuleCopyWith<$Res> get rule;
+}
+
+/// @nodoc
+class _$RecommendedRuleSeverityCopyWithImpl<$Res,
+        $Val extends RecommendedRuleSeverity>
+    implements $RecommendedRuleSeverityCopyWith<$Res> {
+  _$RecommendedRuleSeverityCopyWithImpl(this._value, this._then);
+
+  // ignore: unused_field
+  final $Val _value;
+  // ignore: unused_field
+  final $Res Function($Val) _then;
+
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? rule = null,
+    Object? reason = null,
+    Object? severityLevel = null,
+  }) {
+    return _then(_value.copyWith(
+      rule: null == rule
+          ? _value.rule
+          : rule // ignore: cast_nullable_to_non_nullable
+              as Rule,
+      reason: null == reason
+          ? _value.reason
+          : reason // ignore: cast_nullable_to_non_nullable
+              as String,
+      severityLevel: null == severityLevel
+          ? _value.severityLevel
+          : severityLevel // ignore: cast_nullable_to_non_nullable
+              as SeverityLevel,
+    ) as $Val);
+  }
+
+  @override
+  @pragma('vm:prefer-inline')
+  $RuleCopyWith<$Res> get rule {
+    return $RuleCopyWith<$Res>(_value.rule, (value) {
+      return _then(_value.copyWith(rule: value) as $Val);
+    });
+  }
+}
+
+/// @nodoc
+abstract class _$$RecommendedRuleSeverityDartCopyWith<$Res>
+    implements $RecommendedRuleSeverityCopyWith<$Res> {
+  factory _$$RecommendedRuleSeverityDartCopyWith(
+          _$RecommendedRuleSeverityDart value,
+          $Res Function(_$RecommendedRuleSeverityDart) then) =
+      __$$RecommendedRuleSeverityDartCopyWithImpl<$Res>;
+  @override
+  @useResult
+  $Res call({Rule rule, String reason, SeverityLevel severityLevel});
+
+  @override
+  $RuleCopyWith<$Res> get rule;
+}
+
+/// @nodoc
+class __$$RecommendedRuleSeverityDartCopyWithImpl<$Res>
+    extends _$RecommendedRuleSeverityCopyWithImpl<$Res,
+        _$RecommendedRuleSeverityDart>
+    implements _$$RecommendedRuleSeverityDartCopyWith<$Res> {
+  __$$RecommendedRuleSeverityDartCopyWithImpl(
+      _$RecommendedRuleSeverityDart _value,
+      $Res Function(_$RecommendedRuleSeverityDart) _then)
+      : super(_value, _then);
+
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? rule = null,
+    Object? reason = null,
+    Object? severityLevel = null,
+  }) {
+    return _then(_$RecommendedRuleSeverityDart(
+      rule: null == rule
+          ? _value.rule
+          : rule // ignore: cast_nullable_to_non_nullable
+              as Rule,
+      reason: null == reason
+          ? _value.reason
+          : reason // ignore: cast_nullable_to_non_nullable
+              as String,
+      severityLevel: null == severityLevel
+          ? _value.severityLevel
+          : severityLevel // ignore: cast_nullable_to_non_nullable
+              as SeverityLevel,
+    ));
+  }
+}
+
+/// @nodoc
+
+class _$RecommendedRuleSeverityDart implements RecommendedRuleSeverityDart {
+  const _$RecommendedRuleSeverityDart(
+      {required this.rule, required this.reason, required this.severityLevel});
+
+  @override
+  final Rule rule;
+  @override
+  final String reason;
+  @override
+  final SeverityLevel severityLevel;
+
+  @override
+  String toString() {
+    return 'RecommendedRuleSeverity.dart(rule: $rule, reason: $reason, severityLevel: $severityLevel)';
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$RecommendedRuleSeverityDart &&
+            (identical(other.rule, rule) || other.rule == rule) &&
+            (identical(other.reason, reason) || other.reason == reason) &&
+            (identical(other.severityLevel, severityLevel) ||
+                other.severityLevel == severityLevel));
+  }
+
+  @override
+  int get hashCode => Object.hash(runtimeType, rule, reason, severityLevel);
+
+  @JsonKey(ignore: true)
+  @override
+  @pragma('vm:prefer-inline')
+  _$$RecommendedRuleSeverityDartCopyWith<_$RecommendedRuleSeverityDart>
+      get copyWith => __$$RecommendedRuleSeverityDartCopyWithImpl<
+          _$RecommendedRuleSeverityDart>(this, _$identity);
+
+  @override
+  @optionalTypeArgs
+  TResult when<TResult extends Object?>({
+    required TResult Function(
+            Rule rule, String reason, SeverityLevel severityLevel)
+        dart,
+    required TResult Function(
+            Rule rule, String reason, SeverityLevel severityLevel)
+        flutter,
+  }) {
+    return dart(rule, reason, severityLevel);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult? Function(Rule rule, String reason, SeverityLevel severityLevel)?
+        dart,
+    TResult? Function(Rule rule, String reason, SeverityLevel severityLevel)?
+        flutter,
+  }) {
+    return dart?.call(rule, reason, severityLevel);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function(Rule rule, String reason, SeverityLevel severityLevel)?
+        dart,
+    TResult Function(Rule rule, String reason, SeverityLevel severityLevel)?
+        flutter,
+    required TResult orElse(),
+  }) {
+    if (dart != null) {
+      return dart(rule, reason, severityLevel);
+    }
+    return orElse();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult map<TResult extends Object?>({
+    required TResult Function(RecommendedRuleSeverityDart value) dart,
+    required TResult Function(RecommendedRuleSeverityFlutter value) flutter,
+  }) {
+    return dart(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult? Function(RecommendedRuleSeverityDart value)? dart,
+    TResult? Function(RecommendedRuleSeverityFlutter value)? flutter,
+  }) {
+    return dart?.call(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(RecommendedRuleSeverityDart value)? dart,
+    TResult Function(RecommendedRuleSeverityFlutter value)? flutter,
+    required TResult orElse(),
+  }) {
+    if (dart != null) {
+      return dart(this);
+    }
+    return orElse();
+  }
+}
+
+abstract class RecommendedRuleSeverityDart implements RecommendedRuleSeverity {
+  const factory RecommendedRuleSeverityDart(
+          {required final Rule rule,
+          required final String reason,
+          required final SeverityLevel severityLevel}) =
+      _$RecommendedRuleSeverityDart;
+
+  @override
+  Rule get rule;
+  @override
+  String get reason;
+  @override
+  SeverityLevel get severityLevel;
+  @override
+  @JsonKey(ignore: true)
+  _$$RecommendedRuleSeverityDartCopyWith<_$RecommendedRuleSeverityDart>
+      get copyWith => throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class _$$RecommendedRuleSeverityFlutterCopyWith<$Res>
+    implements $RecommendedRuleSeverityCopyWith<$Res> {
+  factory _$$RecommendedRuleSeverityFlutterCopyWith(
+          _$RecommendedRuleSeverityFlutter value,
+          $Res Function(_$RecommendedRuleSeverityFlutter) then) =
+      __$$RecommendedRuleSeverityFlutterCopyWithImpl<$Res>;
+  @override
+  @useResult
+  $Res call({Rule rule, String reason, SeverityLevel severityLevel});
+
+  @override
+  $RuleCopyWith<$Res> get rule;
+}
+
+/// @nodoc
+class __$$RecommendedRuleSeverityFlutterCopyWithImpl<$Res>
+    extends _$RecommendedRuleSeverityCopyWithImpl<$Res,
+        _$RecommendedRuleSeverityFlutter>
+    implements _$$RecommendedRuleSeverityFlutterCopyWith<$Res> {
+  __$$RecommendedRuleSeverityFlutterCopyWithImpl(
+      _$RecommendedRuleSeverityFlutter _value,
+      $Res Function(_$RecommendedRuleSeverityFlutter) _then)
+      : super(_value, _then);
+
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? rule = null,
+    Object? reason = null,
+    Object? severityLevel = null,
+  }) {
+    return _then(_$RecommendedRuleSeverityFlutter(
+      rule: null == rule
+          ? _value.rule
+          : rule // ignore: cast_nullable_to_non_nullable
+              as Rule,
+      reason: null == reason
+          ? _value.reason
+          : reason // ignore: cast_nullable_to_non_nullable
+              as String,
+      severityLevel: null == severityLevel
+          ? _value.severityLevel
+          : severityLevel // ignore: cast_nullable_to_non_nullable
+              as SeverityLevel,
+    ));
+  }
+}
+
+/// @nodoc
+
+class _$RecommendedRuleSeverityFlutter
+    implements RecommendedRuleSeverityFlutter {
+  const _$RecommendedRuleSeverityFlutter(
+      {required this.rule, required this.reason, required this.severityLevel});
+
+  @override
+  final Rule rule;
+  @override
+  final String reason;
+  @override
+  final SeverityLevel severityLevel;
+
+  @override
+  String toString() {
+    return 'RecommendedRuleSeverity.flutter(rule: $rule, reason: $reason, severityLevel: $severityLevel)';
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$RecommendedRuleSeverityFlutter &&
+            (identical(other.rule, rule) || other.rule == rule) &&
+            (identical(other.reason, reason) || other.reason == reason) &&
+            (identical(other.severityLevel, severityLevel) ||
+                other.severityLevel == severityLevel));
+  }
+
+  @override
+  int get hashCode => Object.hash(runtimeType, rule, reason, severityLevel);
+
+  @JsonKey(ignore: true)
+  @override
+  @pragma('vm:prefer-inline')
+  _$$RecommendedRuleSeverityFlutterCopyWith<_$RecommendedRuleSeverityFlutter>
+      get copyWith => __$$RecommendedRuleSeverityFlutterCopyWithImpl<
+          _$RecommendedRuleSeverityFlutter>(this, _$identity);
+
+  @override
+  @optionalTypeArgs
+  TResult when<TResult extends Object?>({
+    required TResult Function(
+            Rule rule, String reason, SeverityLevel severityLevel)
+        dart,
+    required TResult Function(
+            Rule rule, String reason, SeverityLevel severityLevel)
+        flutter,
+  }) {
+    return flutter(rule, reason, severityLevel);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult? Function(Rule rule, String reason, SeverityLevel severityLevel)?
+        dart,
+    TResult? Function(Rule rule, String reason, SeverityLevel severityLevel)?
+        flutter,
+  }) {
+    return flutter?.call(rule, reason, severityLevel);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function(Rule rule, String reason, SeverityLevel severityLevel)?
+        dart,
+    TResult Function(Rule rule, String reason, SeverityLevel severityLevel)?
+        flutter,
+    required TResult orElse(),
+  }) {
+    if (flutter != null) {
+      return flutter(rule, reason, severityLevel);
+    }
+    return orElse();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult map<TResult extends Object?>({
+    required TResult Function(RecommendedRuleSeverityDart value) dart,
+    required TResult Function(RecommendedRuleSeverityFlutter value) flutter,
+  }) {
+    return flutter(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult? Function(RecommendedRuleSeverityDart value)? dart,
+    TResult? Function(RecommendedRuleSeverityFlutter value)? flutter,
+  }) {
+    return flutter?.call(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(RecommendedRuleSeverityDart value)? dart,
+    TResult Function(RecommendedRuleSeverityFlutter value)? flutter,
+    required TResult orElse(),
+  }) {
+    if (flutter != null) {
+      return flutter(this);
+    }
+    return orElse();
+  }
+}
+
+abstract class RecommendedRuleSeverityFlutter
+    implements RecommendedRuleSeverity {
+  const factory RecommendedRuleSeverityFlutter(
+          {required final Rule rule,
+          required final String reason,
+          required final SeverityLevel severityLevel}) =
+      _$RecommendedRuleSeverityFlutter;
+
+  @override
+  Rule get rule;
+  @override
+  String get reason;
+  @override
+  SeverityLevel get severityLevel;
+  @override
+  @JsonKey(ignore: true)
+  _$$RecommendedRuleSeverityFlutterCopyWith<_$RecommendedRuleSeverityFlutter>
+      get copyWith => throw _privateConstructorUsedError;
+}