+++ tools/update_lint_rules/test/fakes/services/fake_analysis_options_service.dart
@@ -3,6 +3,7 @@ import 'package:update_lint_rules/src/models/dart_sdk_release.dart';
 import 'package:update_lint_rules/src/models/flutter_sdk_release.dart';
 import 'package:update_lint_rules/src/models/lint_rule.dart';
 import 'package:update_lint_rules/src/models/not_recommended_rule.dart';
+import 'package:update_lint_rules/src/models/recommended_rule_severity.dart';
 import 'package:update_lint_rules/src/services/analysis_options_service.dart';
 
 class FakeAnalysisOptionsService implements AnalysisOptionsService {
@@ -11,6 +12,7 @@ class FakeAnalysisOptionsService implements AnalysisOptionsService {
     required Iterable<DartSdkRelease> releases,
     required Iterable<DartLintRule> lintRules,
     required Iterable<NotRecommendedRule> notRecommendedRules,
+    required Iterable<RecommendedRuleSeverity> recommendedRuleSeverities,
   }) async {}
 
   @override
@@ -18,6 +20,7 @@ class FakeAnalysisOptionsService implements AnalysisOptionsService {
     required Iterable<FlutterSdkRelease> releases,
     required Iterable<FlutterLintRule> lintRules,
     required Iterable<NotRecommendedRule> notRecommendedRules,
+    required Iterable<RecommendedRuleSeverity> recommendedRuleSeverities,
   }) async {}
 
   @override
@@ -31,6 +34,7 @@ class FakeAnalysisOptionsService implements AnalysisOptionsService {
   Future<void> outputRecommendedLintRules({
     required File outputFile,
     required Iterable<NotRecommendedRule> notRecommendedRules,
+    required Iterable<RecommendedRuleSeverity> recommendedRuleSeverities,
     required String includeContent,
   }) async {}
 }