+++ tools/update_lint_rules/test/fakes/services/fake_lint_rule_service.dart
@@ -1,5 +1,6 @@
 import 'package:update_lint_rules/src/models/lint_rule.dart';
 import 'package:update_lint_rules/src/models/not_recommended_rule.dart';
+import 'package:update_lint_rules/src/models/recommended_rule_severity.dart';
 import 'package:update_lint_rules/src/services/lint_rule_service.dart';
 
 class FakeLintRuleService implements LintRuleService {
@@ -13,6 +14,14 @@ class FakeLintRuleService implements LintRuleService {
     return (
       dart: <NotRecommendedDartRule>[],
       flutter: <NotRecommendedFlutterRule>[],
+    );
+  }
+
+  @override
+  Future<RecommendedRuleSeverities> getRecommendedRuleSeverities() async {
+    return (
+      dart: <RecommendedRuleSeverityDart>[],
+      flutter: <RecommendedRuleSeverityFlutter>[],
     );
   }